requirements.txt
# Backend core
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
cryptography
httpx
boto3
pandas
scikit-learn
joblib
python-multipart
apscheduler
fastapi-utils
celery[redis]
redis
alembic
requests

# Auth & Security
authlib
itsdangerous
pyjwt
